Refactor tests: Move github, figma, and examples tests to connector artifacts/test folders (#66)
* Initial plan

* Move github, figma, and examples tests to connector artifacts/test folders


* Remove debug console.log statement from FindProjects.test.js


---------
