+++ package.json
@@ -2,8 +2,8 @@
     "name": "appmixer-connectors",
     "scripts": {
         "lint": "eslint . --ext .js --no-error-on-unmatched-pattern",
-        "test-unit": "mocha --recursive --exit --exclude '**/node_modules/**' 'test/**/*.test.js' 'src/appmixer/**/artifacts/test/**/*.test.js'",
-        "test-unit-coverage": "nyc --reporter=lcov --reporter=text-summary mocha --recursive --exit --exclude '**/node_modules/**' 'test/**/*.test.js' 'src/appmixer/**/artifacts/test/**/*.test.js'"
+        "test-unit": "mocha --recursive --exit --exclude '**/node_modules/**' 'test/**/*.test.js' 'src/appmixer/**/artifacts/test/**/*.test.js' 'src/examples/**/artifacts/test/**/*.test.js'",
+        "test-unit-coverage": "nyc --reporter=lcov --reporter=text-summary mocha --recursive --exit --exclude '**/node_modules/**' 'test/**/*.test.js' 'src/appmixer/**/artifacts/test/**/*.test.js' 'src/examples/**/artifacts/test/**/*.test.js'"
     },
     "devDependencies": {
         "autodetect-decoder-stream": "^2.0.2",
