+++ src/appmixer/github/artifacts/test/FindProjects.test.js
@@ -4,10 +4,10 @@ const dotenv = require('dotenv');
 const axios = require('axios');
 
 // Load environment variables
-dotenv.config({ path: path.join(__dirname, '../.env') });
+dotenv.config({ path: path.join(__dirname, '../../../../../test/.env') });
 
 // Import the component
-const FindProjects = require('../../src/appmixer/github/project/FindProjects/FindProjects');
+const FindProjects = require('../../project/FindProjects/FindProjects');
 
 // Mock context
 const createMockContext = (auth, messages = {}) => {
@@ -114,8 +114,6 @@ describe('FindProjects', () => {
         try {
             const result = await FindProjects.receive(context);
 
-            console.log('lasjdflkjdsklfj');
-            console.log(result.data);
             if (result) {
                 assert(result, 'Should return result');
                 assert(result.port === 'out', 'Should use out port');
