+++ src/appmixer/github/artifacts/test/GetProjectFields.test.js
@@ -4,10 +4,10 @@ const dotenv = require('dotenv');
 const axios = require('axios');
 
 // Load environment variables
-dotenv.config({ path: path.join(__dirname, '../.env') });
+dotenv.config({ path: path.join(__dirname, '../../../../../test/.env') });
 
 // Import the component
-const GetProjectFields = require('../../src/appmixer/github/project/GetProjectFields/GetProjectFields');
+const GetProjectFields = require('../../project/GetProjectFields/GetProjectFields');
 
 // Mock context
 const createMockContext = (auth, messages = {}) => {
